solutions_template/requirements.txt
copier
PyYAML
typer
jinja2
jinja2-time
copier-templates-extensions
pathlib